system-install: Add common install target for common config
version.cfg is needed by both client and server, in order to cater to
this the simplest solution seems to be to stuff the whole config dir in
a common install target (instead of picking out the version.cfg file
specifically, and mess with exclusion when handling the config dir in
other targets). In distro packaging, this might make a separate package.