system-install: Add doc link for client finding guidelines.txt
Since the client needs to read guidelines.txt the first time the user
connects to the master server, it needs to be installed in an accessible
location for the client, add a symlink for this purpose.

Also re-arrange so that the install target which has the real data also
takes care of the link, avoiding broken symlinks.