add experimental loadout menu for testing
Replace the grid of buttons with a list of weapons that can be sorted.
Experimental: Although two random slots are shown here, one of them uses weapon id 1, which will actually be ignored and results in use of the next weapon in the list. So, this is to test and discuss a different idea for the loadout buttons in the profile menu.