cannon -> blaster
While the plasma isn't a cannon, it also isn't much of a revolver, nor can it use the name pistol as that is in Halo and is already used by one of the other weapons. That's why I propose with this Propose file change button we call the plasma a plasma blaster.