fix x-y-z entinfo on particle labels
I just had a first look at this, so i might be wrong, but I think the axis of the particle direction was not shown correctly for other values than 256 to 256+32. Not sure what your convention on the coordinate system is, though, but this gives at least the right "periodicity", you agree?