Make bots use weapons except pistol/botweaponspawn (#591)
I'm not sure if it's the right fix for that. That check makes bots spawn with `botweaponspawn` in basic, which is pistol by default, and they just don't want to switch from that (the result is that all bots use pistol). If `botweaponspawn` is set to some other value bots might choose to switch it, however for some weapons like smg and zapper they prefer to keep it. That is why I'm not sure whether something should be changed in how AI switches weapons, or that check should be removed. Or maybe it was meant as `!m_classic`, `!m_basic` doesn't quite make sense to me :)