Support redeclipse:// URLs as command line arguments
This commit introduces the Red Eclipse URL scheme that can be used to start
Red Eclipse and immediately connect to a Red Eclipse server without having to
type a connect command or using the server browser to find a specific server.

The Red Eclipse URL scheme is defined as follows:

redeclipse://password@hostname:port
redeclipse://hostname:port
redeclipse://hostname

Slashes appended to the URL are ignored, hostname and port arguments are
optional.

The parser uses the same functionality as the /connect command, which means
it provides similar behavior.

Beware you need to configure your system to launch the Red Eclipse client
binary when redeclipse:// URLs are opened (e.g. in a Browser or a desktop
shortcut). You can find an example for Linux desktops in the community
repository (https://github.com/red-eclipse/community).