Improve IRC and log privacy
Until now, a user's system's username and hostname have been written to the
log files and were used as identification for the in-game IRC client.

Both is a privacy issue. Being in the log, if an unexperienced user pastes
their log on the internet, both information is shared with third parties.
Using it for the IRC client is even more dangerous, because if some user
accidentally pressed I ingame, the username would be leaked automatically.
As the user is not even informed about this, it might even be illegal in
several countries.

As a result, systemuser and systemhost variables have been removed as they
would no longer be used in the codebase.

The issue has previously been discussed on the forums more than a year ago
(http://redeclipse.net/forum/viewtopic.php?f=7&t=814), but resulted in some
"might be useful". Apparently, this is not an excuse for creating potential
privacy issues.