Move Git repository information code to separate file (implements #663)
This commit moves the code depending on Git revisions to a separate C++ source file that is touched on every build to ensure it is recompiled on every build. If there are modifications, it will cause other files that depend on it to be updated, too. This is maybe not the best solution, but the sanest thing we've come up with on IRC so far.