Update captureresetfloor to proper default value (1).
Currently the default is set to 2, but this means that if I toss my enemies' flag, it resets to my local position. The original post I made addressing this wasn't particularly clear (and was mostly about discussion), but my intention was that wouldn't be the case, as I could potentially toss near death to make it harder for the enemy team to get it back again (but, if a team mate is in the right location, much easier for them).

Also added a comment explaining what the values actually are.