Increase the client/server update rate to 125 Hz and 100 Hz
This decreases effective latency significantly by making clients
and servers sending and receiving updates sooner. This increases
network traffic by a factor of 3, but it should be manageable
on today's networks.

The average game-induced latency (on top of network lag) should
now be 8 ms instead of 40 ms.

See https://extra-a.github.io/#bnet for more information.