generalize architecture handling a little
Now ppc and aarch64 systems are properly detected and used in both
the build system and the launcher, there ismore reliable detection
for the makefile via a shell script (using compiler triples, which
match across systems) and so on.

It's not perfect, but better what we have right now.

Also, do not set -ffast-math by default on platforms other than
x86 where it's confirmed safe. On other targets it may generate
unsafe code that will not behave correctly.