passthroughentcancel variable to disable entcancel on passthrough
passthroughentcancel governs canceling of entity selection
when engaging passthrough mode. Default is 0, which changes the old behavior.
Switch to 1 to revert to old behavior.

Change motivated by the need to use the CTRL (and other) key modifiers
in edit mode, thus all actions bound by default to modifier keys should
not alter the editor state (entcancel in this case).