Automatically adjust field of view for widescreen aspect ratios
The FOV value is now always interpreted as being a 4:3 horizontal FOV
value. It will be automatically scaled for different aspect ratios
(which results in a horizontal FOV of ~106.26 degrees on 16:9,
for example).

This provides a better out of the box experience for most players,
since 16:9 is the norm for gaming nowadays (or even wider on mobile).

Default FOV values were changed to be less extreme at wider aspect ratios.