Use in-out easing for better animation interpolation
The difference is mainly noticeable on high refresh-rate monitors,
but it affects both third-person animations and first-person weapons.
The effect on first-person weapons is mainly visible after switching
to certain weapons such as the pistol, shotgun and zapper.

The animation interpolation duration has been slightly lengthened
to better make use of the new easing.