script: add a rudimentary script macro preprocessor
By using the macro preprocessor it's possible to avoid
having to use "@" macro expanders in deep nesting levels.
It will automatically measure current nesting level
and replace all instances of "#" with
an appropriate amount of macro expanders.

Additionally, a digit can be used directly after "#" to
change the level offset, in case you wish to expand
a variable that is not within the global context level.
Note that in such a case, you will still be bound by
variables needing to be resolved during runtime,
as the preprocessor does not resolve the
values of the macros.

Example:

```
    function = [ result "value" ]

    \# foobar [
        local variable
        variable = 420

        if (= $arg1 1) [
            if (> $arg2 100) [
                if (= $arg3 10) [
                    echo #(function)
                    echo #1variable
                ]
            ]
        ]
    ]
```

Will be equal to:

```
    function = [ result "value" ]

    foobar = [
        local variable
        variable = 420

        if (= $arg1 1) [
            if (> $arg2 100) [
                if (= $arg3 10) [
                    echo value
                    // Will not expand to "420",
                    // as the value is not known
                    // at the time of the preprocess
                    echo @@@variable
                ]
            ]
        ]
    ]
```