Allow `smsize 15` for better shadow map resolution
This results in sharper, more stable shadows that better connect
to the object casting them (with less acne/peter-panning).

This is a demanding setting compared to `/smsize 14`, but modern
high-end GPUs can still render the game smoothly with `/smsize 15`
in most maps. Still, given how demanding this setting is,
it's intentionally not exposed in the graphics options menu.