Update server.cpp to get map from vote starter
if the server does not have the map of the winning vote then the server now gets the map from the player who started the vote, instead of the player who was on the server the longest. The old functionality is used as a fallback in case getting the map from the vote starter fails.
If the server already has the map it continues as normal.